% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.surv2test.R
\name{print.surv2test}
\alias{print.surv2test}
\title{print.surv2test}
\usage{
\method{print}{surv2test}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{digits}{Integer indicating the number of decimal places}

\item{...}{Further arguments ignored in this function}
}
\description{
S3 method for class 'surv2test'
}
