% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv2test.R
\name{surv2test}
\alias{surv2test}
\title{Conducting Two Sample Tests with Time-To-Event Data}
\usage{
surv2test(time, status, arm, tau, nmethod=10000, seed=1)
}
\arguments{
\item{time}{The follow-up time for right censored data.}

\item{status}{The status indicator, 1=event, and 0=right censored.}

\item{arm}{The group indicator for comparison. The elements of this vector take either 1 or 0. Normally, 0=control group, 1=active treatment group.}

\item{tau}{A scaler value to specify the specific time point for calculating the t-year event rate difference.
\code{tau} needs to be smaller than the minimum of the largest observed time in each of the two groups. When \code{tau = NULL}, the default value (i.e., the minimum of the largest observed time in each of the two groups) is used.}

\item{nmethod}{A number of iterations for the resampling method. Recommended to specify at least 10000 (default) or larger.}

\item{seed}{An integer value, used for the random number generation in the resampling procedures. Default is \code{NULL}.}
}
\value{
an object of class surv2test.

\item{tau}{the specific time point used for the t-year event rate differnce test}

\item{note}{a note regarding the specific time point}

\item{tyear_pval}{p-value of the t-year event rate difference test}

\item{maxcombo_pval}{p-value of the Max-Combo test}
}
\description{
Performs two-sample comparisons using various tests.
The current version includes the Max-Combo test and the t-year event rate difference test
that were used in Horiguchi et al. (2020) <doi:10.1177/1740774520940256>. All tests are two-sided where the alternative hypothesis is that
treatment group is not equal to control group with respect to survival. For the Max-Combo test, we used a purmutation resampling method to determine the null distribution of the test statistic under the null hypothesis.
}
\examples{
#--- sample data ---#
D=surv2test.sample.data()
time=D$time
status=D$status
arm=D$arm
nmethod = 100 #This is only for example use.
              #Recommended to specify at least 10000 (default) or larger.
a=surv2test(time, status, arm, tau=10, nmethod=nmethod, seed=123)
print(a)
}
\references{
Horiguchi M, Hassett MJ, Uno H. Empirical power comparison of statistical tests in contemporary phase III randomized controlled trials
with time-to-event outcomes in oncology. Clinical Trials 2020. doi:10.1177/1740774520940256.
}
\author{
Miki Horiguchi, Hajime Uno
}
